Factor out common abstract base class
